CloudEngine/requirements.txt
astroid==2.1.0
autopep8==1.4.3
Brotli==1.0.7
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-bootstrap4==0.0.7
django-braces==1.13.0
django-chartit==0.2.9
django-core==1.4.1
django-cors-headers==2.4.0
django-inspectdb-refactor==0.3
django-oauth==1.1
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-user-profile==1.0.3
django-wsgi==1.0b1
dj-static==0.0.6
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.15.4
oauth==1.0.1
oauthlib==2.1.0
openapi-codec==1.3.2
pbr==5.1.1
pep8==1.7.1
pip==10.0.1
pipenv==2018.11.26
pycodestyle==2.4.0
PyJWT==1.7.1
pylint==2.2.2
pyparsing==2.3.0
python3-openid==3.1.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
setuptools==39.0.1
simplejson==3.16.0
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
static3==0.7.0
stevedore==1.30.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
WebOb==1.8.4
whitenoise==4.1.2
wrapt==1.10.11
